fix variable types and default values to match those in OpenCV sources

+++ src/+cv/MSER.cpp
@@ -21,16 +21,16 @@ void mexFunction( int nlhs, mxArray *plhs[],
     // Check the number of arguments
     if (nrhs<1 || ((nrhs%2)!=1) || nlhs>2)
         mexErrMsgIdAndTxt("mexopencv:error","Wrong number of arguments");
-    
+
     // Argument vector
     vector<MxArray> rhs(prhs,prhs+nrhs);
-    
+
     // Option processing
     int _delta = 5;
     int _min_area = 60;
     int _max_area = 14400;
-    float _max_variation = .25f;
-    float _min_diversity = .2f;
+    double _max_variation = .25f;
+    double _min_diversity = .2f;
     int _max_evolution = 200;
     double _area_threshold = 1.01;
     double _min_margin = .003;
@@ -61,7 +61,7 @@ void mexFunction( int nlhs, mxArray *plhs[],
         else
             mexErrMsgIdAndTxt("mexopencv:error","Unrecognized option");
     }
-    
+
     // Process
     MSER mser(_delta, _min_area, _max_area, _max_variation, _min_diversity,
               _max_evolution, _area_threshold, _min_margin, _edge_blur_size);
